created an hook for the token regeneration
