+++ packages/core/admin/admin/src/pages/SettingsPage/pages/ApiTokens/EditView/utils/getDateOfExpiration.js
@@ -2,7 +2,7 @@ import { addDays, format } from 'date-fns';
 import * as locales from 'date-fns/locale';
 
 const getDateOfExpiration = (createdAt, duration, language = 'en') => {
-  if (duration) {
+  if (duration && typeof duration === 'number') {
     const durationInDays = duration / 24 / 60 / 60 / 1000;
 
     return format(addDays(new Date(createdAt), durationInDays), 'PPP', {
